Problem description
There are six input variables for determining how creditworthy potential credit card borrowers are, as well as a separate credit score (the “Eggertopia Data”) for each of a training and test set. For convenience the input variables are provided in both raw and standardized form. 

There are two different types of output variables, for which you should develop two different types of models:

Part 1  - a binary model for predicting “Default” / “No Default” for future credit card applicants; and

Part 2 -  a  numerical forecasting model to estimate expected profit or loss for future loan applicants.  Note that customers who eventually default, or are at high risk of default, can nevertheless be very profitable overall.